Refactor: HomeCollectionViewHandler 생성
HomeViewController에서 CollectionView를 관리하고 업데이트하는 메서드와 프로퍼티들을 HomeCollectionViewHandler class로 이동시켜 뷰컨트롤러를 깔끔하게 유지할 수 있도록 했습니다.
